Add changelog for v0.5.0
Since this is the first release for this repo, I did not fill in an
actual changelog. If we want to (in the future) we could use this file
as an initial feature list... but I don't think it's fully necessary.

Mostly this is so the auto-generated release doesn't have a broken link.

Signed-off-by: Justin Kulikauskas jkulikau@redhat.com